requirements.txt
sqlalchemy~=1.4.45
fastapi~=0.88.0
strawberry-graphql~=0.151.0
uvicorn~=0.20.0
psycopg2~=2.9.5